Reworked ItemListView similar to PropTableView
