+++ src/components/Breadcrumbs.vue
@@ -2,16 +2,17 @@
 import { RouterLink } from "vue-router";
 import { useUiStore } from "@/stores/ui";
 
+const MAX_LENGTH = 50;
+
 const ui = useUiStore();
 </script>
 
 <template>
     <div id="breadcrumbs">
-        <!-- <i class="fa-regular fa-bread-loaf"></i> -->
         <RouterLink to="/" class="breadcrumb"><i class="fa-regular fa-house"></i></RouterLink>
         <span v-if="ui.breadcrumbs.length > 0" class="breadcrumb-separator"><i class="fa-regular fa-chevron-right"></i></span>
         <template v-for="(breadcrumb, index) in ui.breadcrumbs">
-            <RouterLink :to="breadcrumb.url" class="breadcrumb">{{ breadcrumb.name }}</RouterLink>
+            <RouterLink :to="breadcrumb.url" class="breadcrumb">{{ breadcrumb.name.length >= MAX_LENGTH ? breadcrumb.name.slice(0, MAX_LENGTH) + "..." : breadcrumb.name }}</RouterLink>
             <span v-if="index < ui.breadcrumbs.length - 1" class="breadcrumb-separator"><i class="fa-regular fa-chevron-right"></i></span>
         </template>
     </div>
