+++ src/router/index.ts
@@ -26,73 +26,36 @@ const router = createRouter({
             path: "/v/profiles",
             name: "vocprez profiles",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "VocPrez Profiles",
-                class: "prez:VocPrezProfile",
-                content: `A list of <a href="http://www.w3.org/ns/dx/prof/Profile" target="_blank" rel="noopener noreferrer">prof:Profiles</a> used by VocPrez.`
-            }
         },
         {
             path: "/v/vocab",
             name: "vocabs",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "Vocabularies",
-                class: "skos:ConceptScheme",
-                enableSearch: true,
-                content: `A list of <a href="http://www.w3.org/2004/02/skos/core#ConceptScheme" target="_blank" rel="noopener noreferrer">skos:ConceptSchemes</a>.`
-            }
         },
         {
             path: "/v/vocab/:vocabId",
             name: "vocab",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "skos:ConceptScheme",
-                getChildren: true,
-                childPred: "skos:member", // placeholder
-                childTitlePred: "skos:prefLabel",
-                enableSearch: true
-            }
         },
         {
             path: "/v/vocab/:vocabId/:conceptId",
             name: "concept",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "skos:Concept",
-            }
         },
         {
             path: "/v/collection",
             name: "collections",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "Collections",
-                class: "skos:Collection",
-                enableSearch: true,
-                content: `A list of <a href="http://www.w3.org/2004/02/skos/core#Collection" target="_blank" rel="noopener noreferrer">skos:Collections</a>.`
-            }
         },
         {
             path: "/v/collection/:collectionId",
             name: "collection",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "skos:Collection",
-                getChildren: true,
-                childPred: "skos:member",
-                childTitlePred: "skos:prefLabel",
-                childDisplayTitle: "Concepts",
-            }
         },
         {
             path: "/v/collection/:collectionId/:conceptId",
             name: "concept-in-collection",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "ConceptInCollection",
-            }
         },
         {
             path: "/s",
@@ -113,76 +76,36 @@ const router = createRouter({
             path: "/s/profiles",
             name: "spaceprez profiles",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "SpacePrez Profiles",
-                class: "prez:SpacePrezProfile",
-                content: `A list of <a href="http://www.w3.org/ns/dx/prof/Profile" target="_blank" rel="noopener noreferrer">prof:Profiles</a> used by SpacePrez.`
-            }
         },
         {
             path: "/s/datasets",
             name: "datasets",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "Datasets",
-                class: "dcat:Dataset",
-                childButton: { name: "Collections", url: "/collections" },
-                enableSearch: true,
-                content: `A list of <a href="http://www.w3.org/ns/dcat#Dataset" target="_blank" rel="noopener noreferrer">dcat:Datasets</a>.`
-            }
         },
         {
             path: "/s/datasets/:datasetId",
             name: "dataset",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "dcat:Dataset",
-                childPred: "rdfs:member",
-                childButton: { name: "Collections", url: "/collections" },
-                enableSearch: true
-            }
         },
         {
             path: "/s/datasets/:datasetId/collections",
             name: "feature collections",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "Feature Collections",
-                parentType: "dcat:Dataset",
-                childButton: { name: "Features", url: "/items" },
-                enableSearch: true,
-                content: `A list of <a href="http://www.opengis.net/ont/geosparql#FeatureCollection" target="_blank" rel="noopener noreferrer">geo:FeatureCollections</a>.`
-            }
         },
         {
             path: "/s/datasets/:datasetId/collections/:featureCollectionId",
             name: "feature collection",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "geo:FeatureCollection",
-                childPred: "rdfs:member",
-                childButton: { name: "Features", url: "/items" },
-                enableSearch: true
-            }
         },
         {
             path: "/s/datasets/:datasetId/collections/:featureCollectionId/items",
             name: "features",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "Features",
-                parentType: "geo:FeatureCollection",
-                enableSearch: true,
-                content: `A list of <a href="http://www.opengis.net/ont/geosparql#Feature" target="_blank" rel="noopener noreferrer">geo:Features</a>.`
-            }
         },
         {
             path: "/s/datasets/:datasetId/collections/:featureCollectionId/items/:featureId",
             name: "feature",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "geo:Feature",
-            }
         },
         {
             path: "/c",
@@ -198,43 +121,21 @@ const router = createRouter({
             path: "/c/profiles",
             name: "catprez profiles",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "CatPrez Profiles",
-                class: "prez:CatPrezProfile",
-                content: `A list of <a href="http://www.w3.org/ns/dx/prof/Profile" target="_blank" rel="noopener noreferrer">prof:Profiles</a> used by CatPrez.`
-            }
         },
         {
             path: "/c/catalogs",
             name: "catalogs",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "Catalogs",
-                class: "dcat:Catalog",
-                enableSearch: true,
-                content: `A list of <a href="http://www.w3.org/ns/dcat#Catalog" target="_blank" rel="noopener noreferrer">dcat:Catalogs</a>.`
-            }
         },
         {
             path: "/c/catalogs/:catalogId",
             name: "catalog",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "dcat:Catalog",
-                getChildren: true,
-                childPred: "dcterms:hasPart",
-                childTitlePred: "rdfs:label",
-                childDisplayTitle: "Resources",
-                enableSearch: true
-            }
         },
         {
             path: "/c/catalogs/:catalogId/:resourceId",
             name: "resource",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "dcat:Resource",
-            }
         },
         {
             path: "/sparql",
@@ -255,19 +156,11 @@ const router = createRouter({
             path: "/profiles",
             name: "profiles",
             component: () => import("@/views/ItemListView.vue"),
-            props: {
-                title: "Profiles",
-                class: "prof:Profile",
-                content: `A list of <a href="http://www.w3.org/ns/dx/prof/Profile" target="_blank" rel="noopener noreferrer">prof:Profiles</a> used across Prez.`
-            }
         },
         {
             path: "/profiles/:profileId",
             name: "profile",
             component: () => import("@/views/PropTableView.vue"),
-            props: {
-                type: "prof:Profile",
-            }
         },
         {
             path: "/object",
