+++ src/views/ObjectView.vue
@@ -2,7 +2,7 @@
 import { onMounted } from "vue";
 import { useRoute } from "vue-router";
 import { useUiStore } from "@/stores/ui";
-import PropTableViewNew from "@/views/PropTableViewNew.vue";
+import PropTableView from "@/views/PropTableView.vue";
 
 const route = useRoute();
 const ui = useUiStore();
@@ -18,7 +18,7 @@ onMounted(() => {
 </script>
 
 <template>
-    <PropTableViewNew v-if="route.query && route.query.uri" />
+    <PropTableView v-if="route.query && route.query.uri" />
     <template v-else>
         <h1 class="page-title">Get Object by URI</h1>
         <p>This page can be used to find objects by their URI (Uniform Resource Identifier) by adding <code>?uri=https://some.uri.here</code> (for example) to the end of this page's URL.</p>
